Feature: New document from template (#16)
Allow switching to edit mode from preview dialog.
Add a "View in Collabora" file action to open the full-screen modal in editing mode.
Create file templates for new files and add an API to create a post with the uploaded file.
Create a modal to create a file using a template.

* Remove unused constant

* Very minimal document templates for Collabora Online

These files contain no data at all. This guarantees that when the
user opens a new document in the app, the language of paragraph,
page size, cell date format, currency, etc. will be according to
the current locale.

* Add file upload icons

* Allow editing file in preview modal

* Modal and API to create a new file